+++ mongeul/src/components/setting/atoms/RadioButton.tsx
@@ -1,33 +1,31 @@
 interface RadioButtonProps {
-  label: string;
   value: string;
   selected: boolean;
-  onClick: () => void;
+  onChange: () => void;
 }
 
 export default function RadioButton({
-  label,
   value,
   selected,
-  onClick,
+  onChange,
 }: RadioButtonProps) {
   return (
-    <label className="w-full flex flex-row justify-between">
+    <div className="flex items-center gap-2 cursor-pointer">
       <input
         type="radio"
         value={value}
         checked={selected}
-        onChange={onClick}
+        onChange={onChange}
         className="hidden"
       />
       <div
         className={`w-5 h-5 rounded-full border-2 ${
           selected ? "bg-theme-500 border-theme-500" : "border-gray-400"
         } flex items-center justify-center`}
+        onClick={onChange}
       >
         {selected && <div className="w-2.5 h-2.5 bg-white rounded-full" />}
       </div>
-      <span>{label}</span>
-    </label>
+    </div>
   );
 }
